==> notebooks/1_Drug_Disease_Preprocessing_Exploratory.ipynb <==
import pyTigerGraph as tg
import pandas as pd
from pandasql import sqldf
pysqldf = lambda q: sqldf(q, globals())

data_dir = '../core/dataset'
pd.options.display.max_colwidth = None

## Load & Format Data

# Stanford BioSnap Datasets

## Disease-drug relations

disease_drug_file = f'{data_dir}/input/DCh-Miner_miner-disease-chemical.tsv'
disease_drug_df = pd.read_csv(disease_drug_file, sep='\t')
# Rename columns
disease_drug_df.rename(columns={'# Disease(MESH)': 'Disease Mesh ID'}, inplace=True)

### Drug attribute information

drug_info_file = f'{data_dir}/input/drugbank vocabulary.csv'
drug_info_df = pd.read_csv(drug_info_file, usecols=['DrugBank ID', 'Common name', 'Synonyms'])
#drug_info_df.fillna('', inplace=True)

### Extract entity-name mappings (Drug)

# Series of common names
drug_com_df = drug_info_df[['DrugBank ID', 'Common name']].rename(columns={'Common name': 'Name'})
drug_com_df = drug_com_df[drug_com_df['Name'].notna()]
drug_com_df['Name'] = drug_com_df['Name'].str.strip()
drug_com_df['isCommon'] = True

# ' | ' delimited str of other names for the drug
drug_syn_df = drug_info_df[['DrugBank ID', 'Synonyms']].rename(columns={'Synonyms': 'Name'})
drug_syn_df['Name'] = drug_syn_df['Name'].str.split("\|")

# Explode the list of other names (excluding common names)
drug_syn_expl_df = drug_syn_df.explode('Name')
drug_syn_expl_df['Name'] = drug_syn_expl_df['Name'].str.strip()
drug_syn_expl_df = drug_syn_expl_df[~drug_syn_expl_df['Name'].isin(drug_com_df['Name'])]
drug_syn_expl_df = drug_syn_expl_df[drug_syn_expl_df['Name'].notna()]
drug_syn_expl_df['isCommon'] = False

drug_all_names_df = pd.concat((drug_com_df, drug_syn_expl_df))

drug_all_names_df

## Disease Info

disease_info_file = f'{data_dir}/input/D-MeshMiner_miner-disease.tsv'
disease_info_df = pd.read_csv(disease_info_file, sep='\t')
disease_info_df.rename(columns={'# MESH_ID': 'Disease Mesh ID', 'Definitions': 'Description'}, inplace=True)
disease_info_df['Description'].fillna('', inplace=True)

disease_info_df.isna().sum(axis=0)

### Extract entity-name mappings (Disease)

# Series of common names
disease_com_df = disease_info_df[['Disease Mesh ID', 'Name']]
disease_com_df = disease_com_df[disease_com_df['Name'].notna()]
disease_com_df['Name'] = disease_com_df['Name'].str.strip()
disease_com_df['isCommon'] = True

# '|' delimited str of other names for the disease
disease_syn_df = disease_info_df[['Disease Mesh ID', 'Synonyms']].rename(columns={'Synonyms': 'Name'})
disease_syn_df['Name'] = disease_syn_df['Name'].str.split("\|")

# Explode the list of other names (excluding common names)
disease_syn_expl_df = disease_syn_df.explode('Name')
disease_syn_expl_df['Name'] = disease_syn_expl_df['Name'].str.strip()
disease_syn_expl_df = disease_syn_expl_df[~disease_syn_expl_df['Name'].isin(disease_com_df['Name'])]
disease_syn_expl_df = disease_syn_expl_df[disease_syn_expl_df['Name'].notna()]
disease_syn_expl_df['isCommon'] = False

disease_all_names_df = pd.concat((disease_com_df, disease_syn_expl_df))
disease_all_names_df

disease_syn_df = disease_info_df[['Disease Mesh ID', 'Synonyms']].rename(columns={'Synonyms': 'Name'})
disease_syn_df['Name'].str.split('\|')

## Write Formatted Data

disease_drug_df.to_csv(f'{data_dir}/output/disease-drug.tsv', index=False, sep='\t')
drug_info_df.to_csv(f'{data_dir}/output/drug.tsv', index=False, sep='\t')
disease_info_df.to_csv(f'{data_dir}/output/disease.tsv', index=False, sep='\t')
drug_all_names_df.to_csv(f'{data_dir}/output/drug-name.tsv', index=False, sep='\t')
disease_all_names_df.to_csv(f'{data_dir}/output/disease-name.tsv', index=False, sep='\t')
