==> notebooks/2_Drug_Diease_Graph_Generator.ipynb <==
import pyTigerGraph as tg
from pyTigerGraph.pyTigerGraph import TigerGraphException
import pandas as pd

# Create & Test Connection

secret = "" # provided on request
cred = {
    "host": "https://gradient-test.i.tgcloud.io",
    "graphname": "gradient",
    "username": "tigergraph",
    "password": "" # provided on request
}


token, token_life_sec, token_expir_date = tg.TigerGraphConnection(**cred).getToken(secret="") # provided on request

cred["apiToken"] = token
conn = tg.TigerGraphConnection(**cred)
print(conn.gsql('ls')) # test connection

# Create Graph

graph_not_exists_message = "Graph name gradient cannot be found. Please provide a valid graph name via 'graph=xxx' parameter string."

try:
    conn.getSchema() # Check if graph exists
    print(conn.gsql('DROP GRAPH gradient'))
except TigerGraphException as e:
    print(e.message)

create_graph_query_fname = "../core/graph/create_graph_schema.gsql"

with open(create_graph_query_fname, "r") as f:
    create_graph_query = f.read()

print(conn.gsql(create_graph_query))

# Load Graph

data_dir = '../core/dataset/output'

drug_df = pd.read_csv(f'{data_dir}/drug.tsv', sep='\t')
drug_df.fillna('', inplace=True)

conn.upsertVertexDataFrame(
    drug_df,
    'Drug',
    v_id = 'DrugBank ID',
    attributes = {
        'common_name': 'Common name'
    }
)

disease_df = pd.read_csv(f'{data_dir}/disease.tsv', sep='\t')
disease_df.fillna('', inplace=True)

conn.upsertVertexDataFrame(
    disease_df,
    'Disease',
    v_id = 'Disease Mesh ID',
    attributes = {
        'common_name': 'Name',
        'description': 'Description'
    }
)

drug_disease_df = pd.read_csv(f'{data_dir}/disease-drug.tsv', sep='\t')

conn.upsertEdgeDataFrame(
    drug_disease_df,
    'Drug',
    'interacts_with',
    'Disease',
    from_id = 'Chemical',
    to_id = 'Disease Mesh ID',
    attributes = {}
)

drug_to_name_df = pd.read_csv(f'{data_dir}/drug-name.tsv', sep='\t')
drug_names_df = drug_to_name_df[['Name', 'isCommon']].drop_duplicates()

print(conn.upsertVertexDataFrame(
    drug_name_df,
    'Name',
    attributes={
        'name': 'Name',
        'isCommon': 'isCommon'
    }
))

print(conn.upsertEdgeDataFrame(
    drug_to_name_df,
    'Drug',
    'has_name',
    'Name',
    attributes={}
))

disease_to_name_df = pd.read_csv(f'{data_dir}/disease-name.tsv', sep='\t')
disease_names_df = disease_to_name_df['Name', 'isCommon'].drop_duplicates()

print(conn.upsertVertexDataFrame(
    disease_name_df,
    'Name',
    attributes={
        'name': 'Name',
        'isCommon': 'isCommon'
    }
))

print(conn.upsertEdgeDataFrame(
    disease_to_name_df,
    'Disease',
    'has_name',
    'Name',
    attributes={}
))
