==> notebooks/3_Drug_Disease_Rating_Prescriber_Preprocessing_Exploratory.ipynb <==
from IPython.display import display
import pandas as pd
from pandasql import sqldf
pysqldf = lambda q: sqldf(q, globals())

data_dir = '../core/dataset'
pd.options.display.max_colwidth = None

## Load & Format Data

# Prescriber Drug Dataset

prescriber_drug_names_df = pd.read_csv(f'{data_dir}/input/Medicare_Part_D_Prescribers_by_Provider_and_Drug_2019.csv')
prescriber_drug_names_df['GE65_Tot_Drug_Cst'].fillna(-1, inplace=True)

prescriber_drug_names_df.isna().sum(axis=0)

# Drug Rating Dataset

drug_ratings_train_df = pd.read_csv(f'{data_dir}/input/drugsComTrain_raw.tsv', sep='\t')
drug_ratings_test_df = pd.read_csv(f'{data_dir}/input/drugsComTest_raw.tsv', sep='\t')

drug_ratings_df = pd.concat((drug_ratings_train_df, drug_ratings_test_df))
drug_ratings_df.drop(['Unnamed: 0', 'usefulCount'], axis=1, inplace=True)
drug_ratings_df.dropna(inplace=True)
drug_ratings_df = drug_ratings_df[~drug_ratings_df['condition'].str.match('\d*</span>')]
drug_ratings_df.iloc[:2]

# Drug Information

### Get all rows for prescriber drugs that have ratings

brnd_name_overlap = prescriber_drug_names_df['Brnd_Name'].isin(set(drug_ratings_df['drugName']))
gnrc_name_overlap = prescriber_drug_names_df['Gnrc_Name'].isin(set(drug_ratings_df['drugName']))

filtered_prescriber_drug_names_df = prescriber_drug_names_df[brnd_name_overlap | gnrc_name_overlap]
filtered_prescriber_drug_names_df.index.name = 'drug_id'

### Get list of unique prescribed drugs

unique_drug_names_df = filtered_prescriber_drug_names_df[['Brnd_Name', 'Gnrc_Name']].drop_duplicates().reset_index()
unique_drug_names_df.drop('drug_id', axis=1, inplace=True)
unique_drug_names_df.index.name = 'drug_id'

unique_drug_names_df

A generic name can have multiple brand names associated with it; we will use generic name as the drug identifier.

unique_drug_names_df[unique_drug_names_df['Gnrc_Name'] == 'Azithromycin']

### Write Formatted Dataset

unique_drug_names_df.to_csv(f'{data_dir}/output/drug-ids.tsv', index=False, sep='\t')

# Prescriber Information

### Extract prescriber entities

prescriber_entities_df = filtered_prescriber_drug_names_df[['Prscrbr_NPI', 'Prscrbr_Last_Org_Name', 'Prscrbr_First_Name']].drop_duplicates()

### Extract Prescriber Type

prescriber_type_df = filtered_prescriber_drug_names_df[['Prscrbr_NPI', 'Prscrbr_Type']]

unique_prescriber_type_df = prescriber_type_df['Prscrbr_Type'].drop_duplicates()

### Extract Prescriber Drug relation

prescriber_drug_cost_edge_df = filtered_prescriber_drug_names_df[['Prscrbr_NPI', 'Gnrc_Name', 'Tot_Drug_Cst', 'GE65_Tot_Drug_Cst']]

filtered_prescriber_drug_names_df.iloc[:3]

### Write Formatted Dataset

prescriber_entities_df.to_csv(f'{data_dir}/output/prescriber.tsv', index=False, sep='\t')
prescriber_type_df.to_csv(f'{data_dir}/output/prescriber-type.tsv', index=False, sep='\t')
unique_prescriber_type_df.to_csv(f'{data_dir}/output/unique_prescriber_type.tsv', index=False, sep='\t')
prescriber_drug_cost_edge_df.to_csv(f'{data_dir}/output/prescriber-drug.tsv', index=False, sep='\t')

prescriber_drug_cost_edge_df

# Disease Info

### Get unique disease/condition names from ratings file

conditions_df = drug_ratings_df['condition'].drop_duplicates()
conditions_df.to_csv(f'{data_dir}/output/disease-names.tsv', index=False, sep='\t')

# Rating/Review Info

ratings_with_gnrc_df = pysqldf("select distinct a.*, b.Gnrc_Name from drug_ratings_df as a join unique_drug_names_df as b on a.drugName = b.Brnd_Name or a.drugName = b.Gnrc_Name")
ratings_with_gnrc_df.index.name = "review_id"
ratings_with_gnrc_df.iloc[:2]

### Get drug-disease (used_for) edges

used_for_df = ratings_with_gnrc_df[['Gnrc_Name', 'condition']].drop_duplicates()

### Get review vertices

reviews_df = ratings_with_gnrc_df[['review', 'rating', 'date']]
reviews_df.index.name = "review_id"

print(reviews_df.shape)
reviews_df.iloc[:2]

### Get review edges

# review drug edge (has_review)
review_drug_df = ratings_with_gnrc_df['Gnrc_Name']

# review disease edge (associated with)
review_disease_df = ratings_with_gnrc_df['condition']

print(review_drug_df)
print(review_disease_df)

### Write Formatted Data

used_for_df.to_csv(f'{data_dir}/output/drug-disease.tsv', index=False, sep='\t')
reviews_df.to_csv(f'{data_dir}/output/review.tsv', index=True, sep='\t')
review_drug_df.to_csv(f'{data_dir}/output/drug-review.tsv', index=True, sep='\t')
review_disease_df.to_csv(f'{data_dir}/output/disease-review.tsv', index=True, sep='\t')

# Exploratory Analysis

prescriber_distinct_df = prescriber_drug_names_df.sample(20000)

prescriber_distinct_df

prescriber_distinct_df = prescriber_distinct_df.replace(-1, 0)

prescriber_distinct_df_1 = pysqldf("select Prscrbr_NPI, Gnrc_Name, avg(Tot_Drug_Cst) as avg_drug_cost, avg(GE65_Tot_drug_Cst) as avg_drug_cost_ge65 from prescriber_distinct_df group by Prscrbr_NPI, Gnrc_Name")

prescriber_distinct_df_1

prescriber_distinct_df_1.to_csv(f"{data_dir}/output/Prescriber_Average_Drug_Costs.tsv", sep='\t')

low_rated_drugs_df = pysqldf("select Gnrc_Name, condition, avg(rating) as avg_rating from ratings_with_gnrc_df where rating <= 5.0 group by Gnrc_Name, condition")

low_rated_drugs_df

joined_prescriber_low_ratings_df = pysqldf("select b.*, a.condition, a.avg_rating from low_rated_drugs_df as a join prescriber_distinct_df_1 as b on lower(a.Gnrc_Name) = lower(b.Gnrc_Name)")

sorted_joined_prescriber_df = pysqldf("select * from joined_prescriber_low_ratings_df order by avg_drug_cost desc")

sorted_joined_prescriber_df

sorted_joined_prescriber_df
