==> notebooks/4_Drug_Disease_Rating_Prescriber_Graph_Generator.ipynb <==
import numpy as np
import pyTigerGraph as tg
from pyTigerGraph.pyTigerGraph import TigerGraphException
import pandas as pd
from tqdm import tqdm

batch_size = 10000

# Create & Test Connection

secret = "e4olhqejp4a24kpbpma4c8hcb9o83iv1"
cred = {
    "host": "https://gradient-test.i.tgcloud.io",
    "graphname": "PrescribedDrugReview",
    "username": "public_access",
    "password": "public_access"
}


token, token_life_sec, token_expir_date = tg.TigerGraphConnection(**cred).getToken(secret=secret)
cred["apiToken"] = token

conn = tg.TigerGraphConnection(**cred)
print(conn.gsql('ls')) # test connection

# Create Graph

graph_not_exists_message = f"Graph name {cred['graphname']} cannot be found. Please provide a valid graph name via 'graph=xxx' parameter string."

try:
    conn.getSchema() # Check if graph exists
    #print(conn.gsql(f"DROP GRAPH {cred['graphname']}"))
except TigerGraphException as e:
    print(e.message)

create_graph_query_fname = f"../core/graph/create_graph_schema_{cred['graphname']}.gsql"

with open(create_graph_query_fname, "r") as f:
    create_graph_query = f.read()

#print(conn.gsql(create_graph_query))

# Load Graph

data_dir = '../core/dataset/output'

### Load Drug & Brand name vertices & relation

drug_df = pd.read_csv(f'{data_dir}/drug-ids.tsv', sep='\t')

print(
    conn.upsertVertexDataFrame(
        drug_df,
        'Drug',
        v_id = 'Gnrc_Name',
        attributes = {}
    )
)

print(
    conn.upsertVertexDataFrame(
        drug_df,
        'BrandName',
        v_id = 'Brnd_Name',
        attributes = {}
    )
)

print(
    conn.upsertEdgeDataFrame(
        drug_df,
        'Drug',
        'has_name',
        'BrandName',
        from_id = 'Gnrc_Name',
        to_id = 'Brnd_Name',
        attributes = {}
    )
)

### Load Prescriber & Type & Prescriber-Drug edges

prescriber_entities_df = pd.read_csv(f'{data_dir}/prescriber.tsv', sep='\t')
prescriber_type_df = pd.read_csv(f'{data_dir}/prescriber-type.tsv', sep='\t')
unique_prescriber_type_df = pd.read_csv(f'{data_dir}/unique_prescriber_type.tsv', sep='\t')
prescriber_drug_cost_edge_df = pd.read_csv(f'{data_dir}/prescriber-drug.tsv', sep='\t')

unique_prescriber_type_df

conn.upsertVertexDataFrame(
    prescriber_entities_df,
    'Prescriber',
    v_id = 'Prscrbr_NPI',
    attributes = {
        'first_name': 'Prscrbr_First_Name',
        'last_name': 'Prscrbr_Last_Org_Name'
    }
)

conn.upsertVertexDataFrame(
    unique_prescriber_type_df,
    'Prescriber_Type',
    v_id = 'Prscrbr_Type',
    attributes = {}
)

num_batches = int(np.ceil(prescriber_type_df.shape[0] / batch_size))
for i in tqdm(range(num_batches)):
    conn.upsertEdgeDataFrame(
        prescriber_type_df[batch_size*i:batch_size*(i+1)],
        'Prescriber',
        'has_prescriber_type',
        'Prescriber_Type',
        from_id = 'Prscrbr_NPI',
        to_id = 'Prscrbr_Type',
        attributes = {}
    )

num_batches = int(np.ceil(prescriber_drug_cost_edge_df.shape[0] / batch_size))
for i in tqdm(range(num_batches)):
    conn.upsertEdgeDataFrame(
        prescriber_drug_cost_edge_df[batch_size*i:batch_size*(i+1)],
        'Prescriber',
        'prescribed',
        'Drug',
        from_id = 'Prscrbr_NPI',
        to_id = 'Gnrc_Name',
        attributes = {
            'total_drug_cost': 'Tot_Drug_Cst',
            'total_drug_cost_ge65': 'GE65_Tot_Drug_Cst'
        }
    )

### Load Disease Vertices

conditions_df = pd.read_csv(f'{data_dir}/disease-names.tsv', sep='\t')
used_for_df = pd.read_csv(f'{data_dir}/drug-disease.tsv', sep='\t')

print(
    conn.upsertVertexDataFrame(
        conditions_df,
        'Disease',
        v_id = 'condition',
        attributes = {}
    )
)

print(
conn.upsertEdgeDataFrame(
    used_for_df,
    'Drug',
    'used_for',
    'Disease',
    from_id = 'Gnrc_Name',
    to_id = 'condition',
    attributes = {}
))

### Load Review Nodes & Edges

reviews_df = pd.read_csv(f'{data_dir}/review.tsv', sep='\t')
review_drug_df = pd.read_csv(f'{data_dir}/drug-review.tsv', sep='\t')
review_disease_df = pd.read_csv(f'{data_dir}/disease-review.tsv', sep='\t')

reviews_df = reviews_df.astype({'rating': int})
reviews_df['date'] = pd.to_datetime(reviews_df['date']).astype(str)
reviews_df.iloc[:2]

num_batches = int(np.ceil(reviews_df.shape[0] / batch_size))
for i in tqdm(range(num_batches)):
    print(conn.upsertVertexDataFrame(
        reviews_df[batch_size*i:batch_size*(i+1)],
        'Review',
        v_id = 'review_id',
        attributes={
            'review': 'review',
            'rating': 'rating',
            'date': 'date'
        }
    ))

num_batches = int(np.ceil(review_drug_df.shape[0] / batch_size))
for i in tqdm(range(num_batches)):
    print(conn.upsertEdgeDataFrame(
        review_drug_df[batch_size*i:batch_size*(i+1)],
        'Drug',
        'has_review',
        'Review',
        from_id = 'Gnrc_Name',
        to_id = 'review_id',
        attributes={}
    ))

num_batches = int(np.ceil(review_disease_df.shape[0] / batch_size))
for i in tqdm(range(num_batches)):
    print(conn.upsertEdgeDataFrame(
        review_disease_df[batch_size*i:batch_size*(i+1)],
        'Review',
        'associated_with',
        'Disease',
        from_id = 'review_id',
        to_id = 'condition',
        attributes={}
    ))
